Tiny refactors to Grackle
- Move bindings into their own scope
- Remove visitor type
